move utils to utils dir + change namespace
