+++ src/plugins/intel_gna/src/transformations/utils/gather_sinking_utils.hpp
